backend/pyproject.toml
[tool.poetry]
name = "booking-htmx"
version = "0.1.0"
description = ""
authors = ["Paul Sheridan <paul.sheridan@me.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic-settings = "^2.2.1"
tenacity = "^8.2.3"
pydantic-extra-types = "^2.6.0"
passlib = "^1.7.4"
alembic = "^1.13.1"
python-jose = "^3.3.0"
python-multipart = "^0.0.9"
sqlalchemy = "^2.0.29"
emails = "^0.6"
jinja2 = "^3.1.3"
mypy = "^1.9.0"
psycopg = {extras = ["binary"], version = "^3.1.18"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
